suggestions and minor tweaks to move to the coordinator pattern.
