+++ OperationSample/Contexts/AuthContext.swift
@@ -8,16 +8,7 @@
 
 import Foundation
 
-class AuthContext {
-    static let shared: AuthContext = {
-        
-        var plistURL = Bundle.main.url(forResource: "OAuthOptions", withExtension: "plist")!
-        let data = try! Data(contentsOf: plistURL)
-        
-        let decoder = PropertyListDecoder()
-        let options = try! decoder.decode(OAuthOptions.self, from: data)
-        return AuthContext(oauthOptions: options)
-    }()
+struct AuthContext {
     
     let options: OAuthOptions
     var tokens: [TokenType:Token] = [:]
